Create home dir for user pptr.
Running npm (ex npm test, or npx {script}) resulted in an error about
not being able to write to log files. Apparently this was introduced at
some point by an npm update. The reson for the error was that the
default location for the loga files is ~/.npm, but since the pptr user
did not have a home directory, the log files could not be created. To
fix this, the pptr user now has a home directory.

https://github.com/npm/cli/issues/4838#issuecomment-1116567499